implement randic and wiener index aswell as tests
